Class created = planet help begginers
Erased everything and added a class called planet I created. Just typed this up when practicing and learning how class' work and all their attributes ect. Added this and plan to continue to add more to this as I understand or add more to the document and learn myself. Anyone else is welcome, especially anyone new to Python or any language, as well as anyone willing to help new learners.