Bash Color Picker License
Anyone can copy and use this, as long is it is not used commercially, and/or kept from any individual from using it as they please. So as long as you have the same expectations and promote the same freedoms, and do not restrict the use, modification or any other legal, 100% open source code.